pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xai-blockchain-framework"
version = "0.1.0"
description = "A multi-dimensional framework for XAI evaluation in blockchain fraud detection"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "David Le Sage Oumbe Fokou", email = "dlofokou@etu.uqac.ca" },
    { name = "Wissam Salhab", email = "wsalhab@etu.uqac.ca" },
    { name = "Fehmi Jaafar", email = "fjaafar@uqac.ca" },
]
keywords = [
    "explainable-ai",
    "blockchain",
    "fraud-detection",
    "evaluation",
    "llm-agents",
    "xai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.11",
    "scikit-learn>=1.4",
    "lightgbm>=4.1",
    "joblib>=1.3",
    "torch>=2.1",
    "torch-geometric>=2.4",
    "networkx>=3.2",
    "shap>=0.44",
    "lime>=0.2.0",
    "captum>=0.7",
    "openai>=1.10",
    "tenacity>=8.2",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "jupyter>=1.0",
    "ipykernel>=6.28",
]

[project.urls]
Homepage = "https://github.com/David4Lesage/xai-eval-blockchain-fraud"
Repository = "https://github.com/David4Lesage/xai-eval-blockchain-fraud"
Issues = "https://github.com/David4Lesage/xai-eval-blockchain-fraud/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"xai_blockchain_framework" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["notebooks", "experiments"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]
